Move secret repo validation into secrets subpackage
