+++ pkg/cmd/secret/list/list.go
@@ -109,7 +109,7 @@ func listRun(opts *ListOptions) error {
 			return err
 		}
 
-		err = cmdutil.ValidateHasOnlyOneRemote(opts.HasRepoOverride, opts.Remotes)
+		err = shared.ValidateHasOnlyOneRemote(opts.HasRepoOverride, opts.Remotes)
 		if err != nil {
 			return err
 		}
